Move the list into the child namespace.
